# Back/docs/swagger.yaml
definitions:
  handlers.ClientResponse:
    properties:
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
...
  types.RegisterClientRequest:
    properties:
      password:
        minLength: 8
        type: string
      role:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  types.UpdateClientRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/clients:
    post:
      consumes:
      - application/json
      description: Create a new client account (admin only)
      parameters:
      - description: Admin authentication token
        in: header
        name: X-Admin-Token
        required: true
        type: string
      - description: Client registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.RegisterClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Client created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ClientResponse'
              type: object
        "400":
          description: 'Bad Request: invalid input'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: 'Unauthorized: missing or invalid admin token'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: 'Conflict: username already exists'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Register client
      tags:
      - Admin
  /admin/clients/{id}:
    delete:
      description: Delete a client account (admin only)
      parameters:
      - description: Admin authentication token
        in: header
        name: X-Admin-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Client deleted successfully
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: 'Bad Request: invalid ID'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: 'Unauthorized: missing or invalid admin token'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: 'Not Found: client not found'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Delete client
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: Update a client's information (admin only)
      parameters:
      - description: Admin authentication token
        in: header
        name: X-Admin-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Client updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ClientResponse'
              type: object
        "400":
          description: 'Bad Request: invalid input'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: 'Unauthorized: missing or invalid admin token'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: 'Not Found: client not found'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: 'Conflict: username already exists'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Update client
      tags:
      - Admin
  /admin/update/data:
    post:
      consumes:
      - multipart/form-data
      description: Accepts a JSON data file upload (admin only), parses it, and returns
        the number of records
      parameters:
      - description: Admin authentication token
        in: header
        name: X-Admin-Token
        required: true
        type: string
      - description: Data file to upload (JSON format)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'Example: {\"count\": 42}'
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
        "400":
          description: 'Bad Request: missing or invalid file'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: 'Unauthorized: missing or invalid admin token'
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Upload data file
      tags:
      - Admin
...
swagger: "2.0"